fix(pois): cast $3/$4 to float8 in VALUES to resolve inconsistent type inference
PostgreSQL inferred conflicting types for $3/$4 because they appeared once
bare in VALUES and once with ::float8 in ST_MakePoint. Adding explicit casts
in both places makes the types consistent and resolves the insert error.

+++ backend/src/services/googlePoiService.ts
@@ -176,7 +176,7 @@ async function upsertPois(places: any[], poiType: string, dept: string, client: 
 
     await client.query(
       `INSERT INTO poi_cache (name, poi_type, lat, lng, geometry, tags, source, fetched_at)
-       VALUES ($1, $2, $3, $4,
+       VALUES ($1, $2, $3::float8, $4::float8,
                ST_SetSRID(ST_MakePoint($4::float8, $3::float8), 4326),
                $5, 'google_places', NOW())`,
       [
